metal-llama: extract nebula glyph JS to standalone file, fix stub
nebula-module.js is the source of truth for the glyph UI — embedded
as a C string at build time via the same CMake pattern as shaders.
Add missing set_param to the renderer stub so Nix sandbox builds link.
